// src/api/api.enum.ts
export enum EnumEntitiesBaseUrl {
	AUTH = '/auth',
	USERS = '/users',
	CATEGORIES = '/categories',
	SUBCATEGORIES = '/subcategories',
	REVIEWS = '/reviews',
	CHARACTERISTICS = '/characteristics',
	PRODUCTS = '/products',
	ORDERS = '/orders',
	STATISCICS = '/statistics'
}

export enum EnumAuthData {
	USER = 'user',
	ACCESS_TOKEN = 'accessToken',
	REFRESH_TOKEN = 'refreshToken'
}

export enum EnumAuthorizationType {
	REGISTER = 'register',
	LOGIN = 'login'
}